# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Allizo Service API",
        "description": "API documentation for Allizo Service platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://allizo.example.com",
            "description": "Production Server"
        },
        {
            "url": "http://alizo.test/",
            "description": "Production API Server"
        }
    ],
    "paths": {
...
        "/api/allizo/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get all services",
                "operationId": "8d62066637942388fb5940b8b12391b9",
                "responses": {
                    "200": {
                        "description": "Services retrieved successfully"
                    },
                    "404": {
                        "description": "No services found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a new service",
                "operationId": "113e5f67a3771c7edb82c4b539e574fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Haircut"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Professional men's haircut"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "description": "Optional image",
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created successfully"
                    },
                    "500": {
                        "description": "Service creation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/allizo/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get a service by ID",
                "operationId": "4d64b39620a27bf311017e3f8571b04f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service retrieved successfully"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update an existing service",
                "operationId": "ddcfa639b21e93012368594776dfabb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New haircut style"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated haircut service"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "image": {
                                        "description": "New optional image",
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service updated successfully"
                    },
                    "500": {
                        "description": "Service update failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a service by ID",
                "operationId": "f7d688403e58c9ec0fae931c19a5a5a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deleted successfully"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentication using Laravel Sanctum",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}